prometheusreceiver: Estimate missing sum for VM histograms
VictoriaMetrics histogram functions like `histogram_avg()` don't rely on
the `_sum` series so in practice users don't always report it.
Instead, VM's `histogram_avg()` computes an approximate average by
assuming the average value of observations in each bucket is the average
of the bucket's start and end boundaries.

The PromQL `histogram_avg()` relies on `_sum` and `_count` to give an
exact average. When the `_sum` series is missing, the average is
reported as zero. To ease the migration from VM to PromQL, this PR fills
in any missing sums in VM histograms with the same estimated sum that VM
would compute.